-update amlSupervisoryBody and wheredoyoulive page
